# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fugue API
  version: 1.0.0
  description: |-
    Authoritative API contract for Fugue.
    Route registration, served OpenAPI artifacts, and generated checks all derive from this file.
servers:
  - url: /
...
paths:
...
  /v1/projects:
    get:
      operationId: listProjects
      summary: List Projects
      tags: [Projects]
      x-fugue-handler: handleListProjects
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdQueryParam"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createProject
      summary: Create Project
      tags: [Projects]
      x-fugue-handler: handleCreateProject
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/projects/{id}:
    patch:
      operationId: patchProject
      summary: Patch Project
      tags: [Projects]
      x-fugue-handler: handlePatchProject
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchProjectRequest"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteProject
      summary: Delete Project
      tags: [Projects]
      x-fugue-handler: handleDeleteProject
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
        - in: query
          name: cascade
          required: false
          description: When true, queue delete operations for every app in the project, remove any remaining unbound backing services, and delete the project automatically once the last resource is gone.
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDeleteResponse"
        '202':
          description: Delete requested and cleanup queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDeleteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/apps/import-github:
    post:
      operationId: importGitHubApp
      summary: Import Git Hub App
      tags: [Apps]
      x-fugue-handler: handleImportGitHubApp
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportGitHubRequest"
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportGitHubResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/operations/{id}:
    get:
      operationId: getOperation
      summary: Get Operation
      tags: [Operations]
      x-fugue-handler: handleGetOperation
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    RuntimeBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    EdgeToken:
      type: apiKey
      in: query
      name: token
    DownloadToken:
      type: apiKey
      in: query
      name: download_token
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      schema:
        type: string
...
    TenantIdQueryParam:
      in: query
      name: tenant_id
      required: false
      schema:
        type: string
...
    IdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
        maxLength: 200
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    StringMap:
      type: object
      additionalProperties:
        type: string
...
    Project:
      type: object
      additionalProperties: false
      required: [id, tenant_id, name, slug, description, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AppFile:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path:
          type: string
        content:
          type: string
        secret:
          type: boolean
        mode:
          type: integer
          format: int32
...
    ImportGitHubPersistentStorageSeedFile:
      type: object
      additionalProperties: false
      required: [path, seed_content, service]
      properties:
        service:
          type: string
        path:
          type: string
        seed_content:
          type: string
...
    ServicePersistentStorageOverrideMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/ServicePersistentStorageOverride"
    AppPersistentStorageSpec:
      type: object
      additionalProperties: false
      properties:
        storage_path:
          type: string
        storage_size:
          type: string
        storage_class_name:
          type: string
        reset_token:
          type: string
        mounts:
          type: array
          items:
            $ref: "#/components/schemas/AppPersistentStorageMount"
...
    AppPostgresSpec:
      type: object
      additionalProperties: false
      properties:
        image:
          type: string
        database:
          type: string
        user:
          type: string
        password:
          type: string
        service_name:
          type: string
        runtime_id:
          type: string
        failover_target_runtime_id:
          type: string
        primary_placement_pending_rebalance:
          type: boolean
        storage_size:
          type: string
        storage_class_name:
          type: string
        instances:
          type: integer
          format: int32
        synchronous_replicas:
          type: integer
          format: int32
        resources:
          $ref: "#/components/schemas/ResourceSpec"
...
    Operation:
      type: object
      additionalProperties: false
      required: [id, tenant_id, type, status, execution_mode, requested_by_type, requested_by_id, app_id, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        type:
          type: string
        status:
          type: string
        execution_mode:
          type: string
        requested_by_type:
          type: string
        requested_by_id:
          type: string
        app_id:
          type: string
        source_runtime_id:
          type: string
        target_runtime_id:
          type: string
        desired_replicas:
          type: integer
          format: int32
        desired_spec:
          $ref: "#/components/schemas/AppSpec"
        desired_source:
          description: Build input requested by the operation.
          $ref: "#/components/schemas/AppSource"
        desired_origin_source:
          description: Durable source ownership that should persist after the operation completes.
          $ref: "#/components/schemas/AppSource"
        result_message:
          type: string
        manifest_path:
          type: string
        assigned_runtime_id:
          type: string
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    ImportProjectRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    ComposeStackSummary:
      type: object
      additionalProperties: false
      required: [compose_path, primary_service, services, warnings, inference_report]
      properties:
        compose_path:
          type: string
        primary_service:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/ImportServiceDetail"
        warnings:
          type: array
          items:
            type: string
        inference_report:
          type: array
          items:
            $ref: "#/components/schemas/TopologyInference"
...
    ProjectListResponse:
      type: object
      additionalProperties: false
      required: [projects]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
    ProjectResponse:
      type: object
      additionalProperties: false
      required: [project]
      properties:
        project:
          $ref: "#/components/schemas/Project"
    ProjectDeleteResponse:
      type: object
      additionalProperties: false
      required: [project, deleted, delete_requested]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        deleted:
          type: boolean
        delete_requested:
          type: boolean
        operations:
          type: array
          items:
            $ref: "#/components/schemas/Operation"
        queued_operations:
          type: integer
          format: int32
        already_deleting_apps:
          type: integer
          format: int32
        deleted_backing_services:
          type: integer
          format: int32
...
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        tenant_id:
          type: string
        name:
          type: string
        description:
          type: string
    PatchProjectRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
...
    ServiceEnvMap:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/StringMap"
...
    ImportGitHubRequest:
      type: object
      additionalProperties: false
      required: [repo_url]
      properties:
        tenant_id:
          type: string
        project_id:
          type: string
        project:
          $ref: "#/components/schemas/ImportProjectRequest"
        repo_url:
          type: string
        repo_visibility:
          type: string
        repo_auth_token:
          type: string
        branch:
          type: string
        source_dir:
          type: string
        name:
          type: string
        description:
          type: string
        build_strategy:
          type: string
        runtime_id:
          type: string
        replicas:
          type: integer
          format: int32
        network_mode:
          $ref: "#/components/schemas/AppNetworkMode"
        service_port:
          type: integer
          format: int32
        dockerfile_path:
          type: string
        build_context_dir:
          type: string
        env:
          $ref: "#/components/schemas/StringMap"
        service_env:
          $ref: "#/components/schemas/ServiceEnvMap"
        service_persistent_storage:
          $ref: "#/components/schemas/ServicePersistentStorageOverrideMap"
        config_content:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/AppFile"
        startup_command:
          type: string
        persistent_storage:
          $ref: "#/components/schemas/AppPersistentStorageSpec"
        persistent_storage_seed_files:
          type: array
          items:
            $ref: "#/components/schemas/ImportGitHubPersistentStorageSeedFile"
        postgres:
          $ref: "#/components/schemas/AppPostgresSpec"
        idempotency_key:
          type: string
        update_existing:
          type: boolean
        delete_missing:
          type: boolean
        dry_run:
          type: boolean
...
    ImportGitHubResponse:
      type: object
      additionalProperties: false
      properties:
        app:
          $ref: "#/components/schemas/App"
        operation:
          $ref: "#/components/schemas/Operation"
        apps:
          type: array
          items:
            $ref: "#/components/schemas/App"
        operations:
          type: array
          items:
            $ref: "#/components/schemas/Operation"
        compose_stack:
          $ref: "#/components/schemas/ComposeStackSummary"
        fugue_manifest:
          $ref: "#/components/schemas/FugueManifestSummary"
        idempotency:
          $ref: "#/components/schemas/ImportGitHubIdempotency"
        request_in_progress:
          type: boolean
        plan:
          $ref: "#/components/schemas/TopologyDeployPlan"
...
    AppNetworkMode:
      type: string
      description: |
        Omit this field to deploy a public service with a managed route.
        Set `internal` to run the app with a cluster-internal Kubernetes Service
        and readiness checks, but without a public route.
        Set `background` to run the app without a public route, Kubernetes Service,
        or readiness port.
      enum: [background, internal]
...
    OperationResponse:
      type: object
      additionalProperties: false
      required: [operation]
      properties:
        operation:
          $ref: "#/components/schemas/Operation"
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"